+++ src/main/java/fr/anisekai/server/events/anime/AnimeEpisodeDurationUpdatedEvent.java
@@ -3,9 +3,9 @@ package fr.anisekai.server.events.anime;
 import fr.anisekai.server.entities.Anime;
 import fr.anisekai.server.events.AnimeUpdatedEvent;
 
-public class AnimeEpisodeDurationUpdatedEvent extends AnimeUpdatedEvent<Long> {
+public class AnimeEpisodeDurationUpdatedEvent extends AnimeUpdatedEvent<Integer> {
 
-    public AnimeEpisodeDurationUpdatedEvent(Object source, Anime entity, Long previous, Long current) {
+    public AnimeEpisodeDurationUpdatedEvent(Object source, Anime entity, Integer previous, Integer current) {
 
         super(source, entity, previous, current);
     }
