+++ src/main/java/fr/anisekai/server/events/anime/AnimeTotalUpdatedEvent.java
@@ -3,9 +3,9 @@ package fr.anisekai.server.events.anime;
 import fr.anisekai.server.entities.Anime;
 import fr.anisekai.server.events.AnimeUpdatedEvent;
 
-public class AnimeTotalUpdatedEvent extends AnimeUpdatedEvent<Long> {
+public class AnimeTotalUpdatedEvent extends AnimeUpdatedEvent<Integer> {
 
-    public AnimeTotalUpdatedEvent(Object source, Anime entity, Long previous, Long current) {
+    public AnimeTotalUpdatedEvent(Object source, Anime entity, Integer previous, Integer current) {
 
         super(source, entity, previous, current);
     }
